% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcolors.R
\name{ggcolors}
\alias{ggcolors}
\title{ggplot2's Default Color Scheme}
\usage{
ggcolors(n)
}
\arguments{
\item{n}{An integer specifying the number of colors.}
}
\value{
Returns a vector of \code{n} hex colors.
}
\description{
Make a palette with \pkg{ggplot2}'s default color scheme.
}
\examples{
n <- 10
ggcolors(n)

plot(
    1:n, 
    pch=16, 
    cex=2, 
    col= ggcolors(n)
)

barplot(
    stats::setNames(5:14, LETTERS[5:14]), 
    col = ggcolors(n), 
    border = ggcolors(n)
)

boxplot(
    x ~ y, 
    data = data.frame(x = rnorm(1000), 
    y = sample(LETTERS[5:14], 1000, TRUE)), 
    border = ggcolors(n),  
    lwd=2
)

pie(
    stats::setNames(5:14, LETTERS[5:14]), 
    col = ggcolors(n), 
    border = ggcolors(n)
)
}
\author{
John Colby
}
\references{
\url{http://stackoverflow.com/a/8197703/1000343}
}

