% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdual_axis.R
\name{ggdual_axis}
\alias{ggdual_axis}
\title{Dual Y-axis for ggplot2}
\usage{
ggdual_axis(lhs, rhs, angle = 270)
}
\arguments{
\item{lhs}{A plot whose y axis shall be on the left hand side.}

\item{rhs}{A plot whose y axis shall be on the right hand side.}

\item{angle}{Angle to rotate y-axis on right hand side.}
}
\value{
Returns an \code{\link[gridExtra]{arrangeGrob}} with extra class
\code{ggdual_axis} that plots by default.  This allows it to be further
combined with other grobs via \code{\link[gridExtra]{grid.arrange}}.
}
\description{
lot dual y-axis for ggplot2 objects.
}
\examples{
p1 <- ggplot(mtcars, aes(mpg, disp)) +
    geom_line(colour = "blue") +
    theme_bw() +
    theme(plot.margin = unit(c(.5, 1, .5, 0), "cm"))

p2 <- ggplot(mtcars, aes(mpg, drat)) +
    geom_line(colour = "red") +
    theme_bw() +
    theme(plot.margin = unit(c(.5, 1, .5, 0), "cm"))

ggdual_axis(lhs = p1, rhs = p2)
}
\references{
\url{http://stackoverflow.com/a/27608585/1000343}
\url{http://stackoverflow.com/a/25699817/1000343}
}

