% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_global_set_color.R
\name{ggplot_global_set_color}
\alias{ggplot_global_set_color}
\title{Globally Set ggplot2 Colors/Fills}
\usage{
ggplot_global_set_color(color = "black", alter.fill = TRUE,
  alter.color = TRUE, reset = TRUE)
}
\arguments{
\item{color}{A single color character string.}

\item{alter.fill}{logical.  If \code{TRUE} fill is globally altered.}

\item{alter.color}{logical.  If \code{TRUE} colour is globally altered.}

\item{reset}{logical.  If colour and fill are reset to black before applying 
new color.}
}
\description{
Globally set the color and fill of ggplot2.
}
\examples{
ggplot_global_set_color("purple")

ggplot(mtcars, aes(factor(cyl))) + geom_bar()
ggplot(mtcars, aes(hp, mpg, group = factor(cyl))) + geom_point()
}
\references{
\url{http://stackoverflow.com/a/21175042/1000343}
}
\seealso{
\code{\link[ggplot2]{update_geom_defaults}}
}

