% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdual_axis.R
\name{plot.ggdual_axis}
\alias{plot.ggdual_axis}
\title{Plots a ggdual_axis Object}
\usage{
\method{plot}{ggdual_axis}(x, ...)
}
\arguments{
\item{x}{The \code{ggdual_axis} object.}

\item{\ldots}{Other arguments passed to \code{\link[gridExtra]{grid.arrange}}.}
}
\description{
Plots a ggdual_axis object.
}

