% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_remover.R
\name{plot.panel_remover}
\alias{plot.panel_remover}
\title{Plots a panel_remover Object}
\usage{
\method{plot}{panel_remover}(x, ...)
}
\arguments{
\item{x}{The \code{panel_remover} object.}

\item{\ldots}{Other arguments passed to \code{\link[gridExtra]{grid.arrange}}.}
}
\description{
Plots a panel_remover object.
}

