% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbalanced_facet_axis.R
\name{plot.unbalanced_facet_axis}
\alias{plot.unbalanced_facet_axis}
\title{Plots an unbalanced_facet_axis object.}
\usage{
\method{plot}{unbalanced_facet_axis}(x, newpage = is.null(viewport),
  viewport = NULL, ...)
}
\arguments{
\item{x}{The unbalanced_facet_axis object}

\item{newpage}{logical.  If \code{TRUE} \code{\link[grid]{grid.newpage}} is called.}

\item{viewport}{logical.  If character \code{\link[grid]{seekViewport}} is
used.  If an object \code{\link[grid]{pushViewport}} is used.  If \code{NULL}
neither viewport is used.}

\item{\ldots}{ignored}
}
\description{
Plots an unbalanced_facet_axis object.
}
