% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{plotflow}
\alias{package-plotflow}
\alias{plotflow}
\alias{plotflow-package}
\title{plotflow: Tools to speed up workflow associated with plotting}
\description{
Plotting functions designed to increase efficiency around plotting.
}

