% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalign.R
\name{print.ggalign}
\alias{print.ggalign}
\title{Prints a ggalign Object}
\usage{
\method{print}{ggalign}(x, ...)
}
\arguments{
\item{x}{The ggalign object.}

\item{\ldots}{ignored.}
}
\description{
Prints a ggalign object.
}

