% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdual_axis.R
\name{print.ggdual_axis}
\alias{print.ggdual_axis}
\title{Prints a ggdual_axis Object}
\usage{
\method{print}{ggdual_axis}(x, ...)
}
\arguments{
\item{x}{The \code{ggdual_axis} object.}

\item{\ldots}{Other arguments passed to \code{\link[gridExtra]{grid.arrange}}.}
}
\description{
Prints a ggdual_axis object.
}
