% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_remover.R
\name{print.panel_remover}
\alias{print.panel_remover}
\title{Prints a panel_remover Object}
\usage{
\method{print}{panel_remover}(x, ...)
}
\arguments{
\item{x}{The \code{panel_remover} object.}

\item{\ldots}{Other arguments passed to \code{\link[gridExtra]{grid.arrange}}.}
}
\description{
Prints a panel_remover object.
}

