% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.unbalanced_facet_axis}
\alias{print.unbalanced_facet_axis}
\title{Prints a unbalanced_facet_axis object.}
\usage{
\method{print}{unbalanced_facet_axis}(x, newpage = is.null(viewport),
  viewport = NULL, ...)
}
\arguments{
\item{x}{The unbalanced_facet_axis object}

\item{newpage}{logical.  If \code{TRUE} \code{\link[grid]{grid.newpage}} is called.}

\item{viewport}{logical.  If character \code{\link[grid]{seekViewport}} is
used.  If an object \code{\link[grid]{pushViewport}} is used.  If \code{NULL}
neither viewport is used.}

\item{\ldots}{ignored}
}
\description{
Prints a unbalanced_facet_axis object.
}

