% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qcolor}
\alias{qcolor}
\title{Plot Base R colors()}
\usage{
qcolor()
}
\description{
Makes a plot of base R colors from \code{\link[grDevices]{colors}}.
}
\examples{
qcolor()
}
\author{
Earl F. Glynn
}
\references{
\url{http://research.stowers-institute.org/efg/R/Color/Chart/ColorChart.R}
}

