\name{qsymbol}
\alias{qsymbol}
\title{Plot ggplot2/base R Symbols}
\usage{
qsymbol(ggplot2 = TRUE)
}
\arguments{
  \item{ggplot2}{logical.  If \code{TRUE}
  \href{http://docs.ggplot2.org/current/}{ggplot2} symbols
  will be plotted.  If \code{FALSE} base graphics symbols
  will be plotted.}
}
\description{
Makes a plot of ggplto2/base R symbols.
}
\examples{
qsymbol()
}

