% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_basic.R
\name{theme_basic}
\alias{theme_basic}
\title{ggplot2 Theme with No Background or Gridlines.}
\usage{
theme_basic(x = FALSE, y = FALSE, plot.box = FALSE)
}
\arguments{
\item{x}{logical.  If \code{TRUE} vertical gridlines are added.}

\item{y}{logical.  If \code{TRUE} horizontal gridlines are added.}

\item{plot.box}{logical.  If \code{TRUE} a full box surrounds the plot area.  If \code{FALSE} only the x and y axis are shown.}
}
\description{
A ggplot2 theme with no background and no gridlines.
}
\note{
Both x and y can not be \code{TRUE}.  Use 
\code{\link[ggplot2]{theme_bw}} instead.
}
\examples{
ggplot(mtcars, aes(factor(cyl))) + geom_bar() + theme_basic()
ggplot(mtcars, aes(factor(cyl))) + geom_bar() + theme_basic(x = TRUE)
ggplot(mtcars, aes(factor(cyl))) + geom_bar() + theme_basic(y = TRUE)
ggplot(mtcars, aes(factor(cyl))) + geom_bar() + theme_basic(x = TRUE, y = TRUE)
}
\seealso{
\code{\link[ggplot2]{theme}}
}
