% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theme_black}
\alias{theme_black}
\title{ggplot2 Theme with No Background or Gridlines.}
\usage{
theme_black(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{The size to use for text.  Various textual components are
scaled off of this value.}

\item{base_family}{The base font family.}
}
\description{
A ggplot2 theme with no background and no gridlines.
}
\examples{
ggplot(mtcars, aes(factor(cyl))) + geom_bar(fill="white") + theme_black()
dat <- data.frame(y = c(austres), time = time(austres))
ggplot(dat, aes(time, y)) + geom_line(color="lightblue", size=1) + theme_black()
ggplot(movies, aes(rating)) + geom_histogram(color="white") + theme_black()

\dontrun{
library(maps)
crimes <- data.frame(state = tolower(rownames(USArrests)), USArrests)
states_map <-map_data("state")

ggplot(crimes, aes(map_id = state)) +
    geom_map(aes(fill = Murder), map = states_map) +
    expand_limits(x = states_map$long, y = states_map$lat) +
    theme_black() +
    scale_fill_gradient(low="grey10", high="white")
}
}
\author{
Jon Lefcheck (\url{http://jonlefcheck.net})
}
\references{
\url{http://jonlefcheck.net/2013/03/11/black-theme-for-ggplot2-2}
}
\seealso{
\code{\link[ggplot2]{theme}}
}

