% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unbalanced_facet_axis}
\alias{unbalanced_facet_axis}
\title{Add Tick Marks to an Unbalanced facet_wrap}
\usage{
unbalanced_facet_axis(ggplot_obj, position = c("up", "down"),
  newpage = is.null(vp), vp = NULL)
}
\arguments{
\item{ggplot_obj}{An unbalanced ggplot2 \code{\link[ggplot2]{facet_wrap}}
object.}

\item{position}{Either \code{"up"} (match unbalanced facet's postition) or
\code{"down"} (along bottom most axis).}

\item{newpage}{logical.  If \code{TRUE} \code{\link[grid]{grid.newpage}} is
used.}

\item{vp}{logical.  If \code{TRUE} a viewport is used.}
}
\description{
Adds the tick marks to an unblanced grouping of
\code{\link[ggplot2]{facet_wrap}} plots.
}
\examples{
set.seed(10)
mtcars[["new"]] <- sample(LETTERS[1:7], nrow(mtcars), TRUE)

library(ggplot2)

unbalanced_facet_axis(ggplot(mtcars, aes(x=mpg, y=hp)) +
    geom_line() +
    facet_wrap(~new, ncol=2))

unbalanced_facet_axis(ggplot(mtcars, aes(x=mpg, y=hp)) +
    geom_line() +
    facet_wrap(~new, ncol=3), "down")
}
\author{
\href{http://stackoverflow.com/users/1320535/julius}{Julius}
(stackoverflow.com)
}
\references{
\url{http://stackoverflow.com/a/13316126/1000343}
}
\keyword{axis}
\keyword{facet_wrap}

