% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yes_no_bar.R
\name{yes_no_bar}
\alias{yes_no_bar}
\title{Two Choice Horizontal Stacked Bar}
\usage{
yes_no_bar(n.yes, n.no = NULL, n = NULL, labs = c("Yes", "No"),
  digits = 1, fill = c("#51A651", "white"), color = "#CAE4CA",
  border = "grey70", size = 10, include.percent = c(TRUE, FALSE))
}
\arguments{
\item{n.yes}{Number yes (integer).  User must supply 2 of the first three
parameters (\code{n.yes}, \code{n.no}, or \code{n}).}

\item{n.no}{Number no (integer).}

\item{n}{Total number (integer).}

\item{labs}{The labels to use for the two choices that correspond to
\code{n.yes} and \code{n.no}.}

\item{digits}{Digits to use in the percents.}

\item{fill}{The bar fill colors}

\item{color}{The text color.}

\item{border}{The border color.}

\item{size}{The text size.}

\item{include.percent}{A logical vector of 1-2 (if 1 the second will be
\code{FALSE}) stating if percents should be included.  The length of 2
corresponds to the 2 choices; \code{n.yes} and \code{n.no}.  If a single value
is given it is recycled.}
}
\value{
Returns a \code{ggplot} object.
}
\description{
Create a two choice horizontal stacked bar chart in the style of base graphics.
Default color scheme makes it easy to throw into slides and posters as an
infographic style plot.
}
\examples{
yes_no_bar(111, 66)
yes_no_bar(111, 66, fill=c("pink", "ivory"), color="purple")
yes_no_bar(55, n = 166, labs = c("True", "False"))
yes_no_bar(2345, 3456, labs = c("Boy", "Girl"), fill=c("lightblue", "pink"),
    color="grey50", include.percent = TRUE)
yes_no_bar(2345, 3456, labs = c("Boy", "Girl"), fill=c("lightblue", "pink"),
    color="grey50", include.percent = FALSE)
yes_no_bar(9999, n = 22166, fill=c("green", "pink"),
    labs = c("Go:", "Stop"), color = "grey30", digits=0)
}
\keyword{bar}
\keyword{infographic}
