% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalign.R
\name{ggalign}
\alias{ggalign}
\title{Align ggplot2 Graph Edges}
\usage{
ggalign(plot1, plot2)
}
\arguments{
\item{plot1}{A ggplot2 plot.}

\item{plot2}{A ggplot2 plot.}
}
\value{
Returns a stacked grid object with left/right edges aligned.
}
\description{
Align ggplot2 graph edges.
}
\examples{
require(ggplot2)
A <- ggplot(CO2, aes(x=Plant)) + geom_bar() + coord_flip() 
B <- ggplot(CO2, aes(x=Type)) + geom_bar() + coord_flip() 
ggalign(A, B)
}
\references{
\url{http://stackoverflow.com/a/13295880/1000343}
}
\author{
Baptiste Auguie
}
\keyword{align}
