% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_rename_margin.R
\name{label_rename_margin}
\alias{label_rename_margin}
\title{Rename the Margins of ggplot2 facet_grid}
\usage{
label_rename_margin(newname = "Total")
}
\arguments{
\item{newname}{The new strip.text name to overwrite (all).}
}
\description{
Convert the default (all) when setting ggplot2's
\code{\link[ggplot2]{facet_grid}} to \code{margins = TRUE}.
}
\examples{
mtcars2 <- mtcars
mtcars2$gear <- factor(mtcars2$gear)
ggplot(mtcars2, aes(cyl)) +
    geom_point(stat="bin", size = 2, binwidth = 2,
        aes(shape = gear), position = "stack") +
    facet_grid(carb ~ gear, margins = TRUE,
        labeller=label_rename_margin("Total"))

ggplot(mtcars2, aes(cyl)) +
    geom_point(stat="bin", size = 2, binwidth = 2,
        aes(shape = gear), position = "stack") +
    facet_grid(carb ~ gear, margins = "gear",
        labeller=label_rename_margin("Total"))
}
\keyword{margin}
