% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotflow-package.R
\docType{package}
\name{plotflow}
\alias{plotflow}
\alias{package-plotflow}
\alias{plotflow-package}
\title{plotflow: Tools to speed up workflow associated with plotting}
\description{
Plotting functions designed to increase efficiency around plotting.
}
