% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbalanced_facet_axis.R
\name{print.unbalanced_facet_axis}
\alias{print.unbalanced_facet_axis}
\title{Prints an unbalanced_facet_axis object.}
\usage{
\method{print}{unbalanced_facet_axis}(x, ...)
}
\arguments{
\item{x}{The unbalanced_facet_axis object}

\item{\ldots}{arguments passed to the plot method}
}
\description{
Calls \code{\link{plot}} with the same arguments.
}
