% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcolors.R
\name{qcolor}
\alias{qcolor}
\title{Plot Base R colors()}
\usage{
qcolor()
}
\description{
Makes a plot of base R colors from \code{\link[grDevices]{colors}}.
}
\examples{
qcolor()
}
\references{
\url{http://research.stowers-institute.org/efg/R/Color/Chart/ColorChart.R}
}
\author{
Earl F. Glynn
}
